fix type

+++ src/xtd_drawing/src/xtd/drawing/bitmap.cpp
@@ -5,7 +5,7 @@ using namespace xtd;
 using namespace xtd::drawing;
 
 
-drawing::color bitmap::get_pixel(size_t x, size_t y) const {
+drawing::color bitmap::get_pixel(int32_t x, int32_t y) const {
   native::argb color;
   native::image::get_pixel(this->handle(), x, y, color);
   return color::from_argb(color.a, color.r, color.g, color.b);
